+++ COMP559W23FP260843175/src/Scene.h
@@ -7,25 +7,24 @@
 
 class Scene {
 	public:
-		double gravity = -9.81;
-		double dt = 1.0 / 120.0;
-		double flipRatio = 0.9;
+		float gravity = -9.81;
+		float dt = 1.0 / 120.0;
+		float flipRatio = 0.9;
 		int numPressureIters = 100;
 		int numParticleIters = 2;
 		int frameNr = 0;
-		double overRelaxation = 1.9;
+		float overRelaxation = 1.9;
 		bool compensateDrift = true;
 		bool separateParticles = true;
-		double obstacleX = 0.0;
-		double obstacleY = 0.0;
-		double obstacleRadius = 0.15;
+		float obstacleX = 0.0;
+		float obstacleY = 0.0;
+		float obstacleRadius = 0.15;
 		bool paused = true;
 		bool showObstacle = true;
-		double obstacleVelX = 0.0;
-		double obstacleVelY = 0.0;
+		float obstacleVelX = 0.0;
+		float obstacleVelY = 0.0;
 		bool showParticles = true;
 		bool showGrid = false;
-		//Simulator* fluid = nullptr;
 		Scene();
 };
